requirements.txt
py-cord
python-dotenv
certifi
tweepy
